+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Update help with new spelling
 - Better handle of deploys when sync is enabled
 - Fix broken output spinners
+- Improved console messages
+- Disable initial compile of crystal functions when starting Local
 
 ## 0.5.12 - Mar 16 2019
 - Add new branding to Local web templates.